use utc to communicate with other components(front/DM)

+++ JEBuilder/WorkflowBuilder/src/main/java/models/JEWorkflow.java
@@ -258,23 +258,26 @@ public class JEWorkflow extends JEObject {
      * Delete a workflow block
      * */
     public void deleteWorkflowBlock(String id) throws InvalidSequenceFlowException, WorkflowBlockNotFound {
-        for (String blockId : allBlocks.get(id).getInflows().values()) {
-            WorkflowBlock block = this.getBlockById(blockId);
-            deleteSequenceFlow(block.getJobEngineElementID(), id);
-        }
-        for (String blockId : allBlocks.get(id).getOutFlows().values()) {
-            WorkflowBlock block = this.getBlockById(blockId);
-            deleteSequenceFlow(id, block.getJobEngineElementID());
-        }
+       if(allBlocks.contains(id))
+       {
+    	   for (String blockId : allBlocks.get(id).getInflows().values()) {
+               WorkflowBlock block = this.getBlockById(blockId);
+               deleteSequenceFlow(block.getJobEngineElementID(), id);
+           }
+           for (String blockId : allBlocks.get(id).getOutFlows().values()) {
+               WorkflowBlock block = this.getBlockById(blockId);
+               deleteSequenceFlow(id, block.getJobEngineElementID());
+           }
 
-        WorkflowBlock b = allBlocks.get(id);
-        if (b == null) {
-            throw new WorkflowBlockNotFound(JEMessages.WORKFLOW_BLOCK_NOT_FOUND);
-        }
-        allBlocks.remove(id);
-        if (allBlocks.size() == 0) workflowStartBlock = null;
-        b = null;
-        status = Status.NOT_BUILT;
+           WorkflowBlock b = allBlocks.get(id);
+           if (b == null) {
+               throw new WorkflowBlockNotFound(JEMessages.WORKFLOW_BLOCK_NOT_FOUND);
+           }
+           allBlocks.remove(id);
+           if (allBlocks.size() == 0) workflowStartBlock = null;
+           b = null;
+           status = Status.NOT_BUILT;
+       }
 
     }
